try a grant model to pass around our expansion trigger boundary & move our models into their own directory
